+++ R/print.R
@@ -4,12 +4,42 @@
 #' Replace the standard print method for functions with one that performs
 #' syntax highlighting, using ANSI colors, if the terminal supports them.
 #'
-#' @docType package
-#' @name prettycode
-#' @export print.function
-NULL
+#' @export
+
+prettycode <- function() {
+  register_s3_method("prettycode", "print", "function", print.function)
+}
+
+register_s3_method <- function(pkg, generic, class, fun = NULL) {
+  stopifnot(is.character(pkg), length(pkg) == 1)
+  envir <- asNamespace(pkg)
+
+  stopifnot(is.character(generic), length(generic) == 1)
+  stopifnot(is.character(class), length(class) == 1)
+  if (is.null(fun)) {
+    fun <- get(paste0(generic, ".", class), envir = parent.frame())
+  }
+  stopifnot(is.function(fun))
+
+
+  if (pkg %in% loadedNamespaces()) {
+    registerS3method(generic, class, fun, envir = envir)
+  }
+
+  # Always register hook in case package is later unloaded & reloaded
+  setHook(
+    packageEvent(pkg, "onLoad"),
+    function(...) {
+      registerS3method(generic, class, fun, envir = envir)
+    }
+  )
+}
 
 #' Print a function with syntax highlighting
+#'
+#' To turn on pretty printing of functions, you need to call
+#' `prettycode::prettycode()`. It might be a good idea to call it
+#' from your `.Rprofile`.
 #'
 #' @param x Function to print.
 #' @param useSource Whether to use the stored source code, if available.